Deleting Second render()
There must have been a merge because this file ended up with two render() sections The first render() is where I added the Material-Ui Component names. Then somehow another render() was inserted that was overiding my edits. I then mistakingling in a hurry did a pull request on the Kenzie project instead of this one. This last edit should be the final one for this project for the hackathon